Add UserData and FileUpload models and
This commit adds the UserData and FileUpload models to the project. The UserData model is used to store user-related data, such as first_name,last_name, nationalid, and email. The FileUpload model is used to store file-related data, such as file name and status.

In addition, this commit fixes an issue with the request library and urllib library. The issue caused requests to fail when trying to connect to certain URLs. The fix involved updating the versions of these libraries and making some changes to the code to ensure compatibility